Added streaming consumer/producer scripts
Added scripts for producers and consumers in kafka, pulsar, and flink.
These scripts all simulate streaming stock data.
These scripts compute the exponential moving average for 50 and 200 days.
These scripts specify what environment variables need to be set in docker.
These scripts are set up to have a single producer, and any number of consumers.
The work will be split up among the consumers.
This requires the consumers to share a single file system.
Disclaimer: I don't have an easy way to run this, so none of this might work... Good luck